requirements.txt
anthropic>=0.86.0
python-dotenv
